Small upload fix
Added tags #selling and #buying on listings titles
