+++ Hermes/View/MyHistory/IHistoryPage.cs
@@ -1,0 +1,14 @@
+﻿using Hermes.Model.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Hermes.View.MyHistory
+{
+    interface IHistoryPage
+    {
+        List<Listing> Listings { set; }
+    }
+}
